models: extract IssueFinder to its own file
